+++ resources/input.txt
@@ -1,1 +1,1 @@
-abc!abc abaa!abb acb abbc!abc
+abc!abc abaa!abb acb abbc!abc!abcabc
